requirements.txt
absl-py==1.4.0
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.3
apturl==0.5.2
async-timeout==4.0.2
attrs==23.1.0
awscli==1.18.69
backcall==0.1.0
bandit==1.6.2
bcrypt==3.1.7
blinker==1.4
botocore==1.16.19
Brlapi==0.7.0
caffeine==2.9.4
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
Click==7.0
cmake==3.26.3
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
diffusers==0.16.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
ewmh==0.1.6
fasteners==0.14.1
filelock==3.0.12
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.2
gitdb2==2.0.6
GitPython==3.0.7
httplib2==0.14.0
huggingface-hub==0.14.1
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.12.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==3.1.2
jmespath==0.9.4
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lit==16.0.5
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe==0.10.0
monotonic==1.5
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.4
netifaces==0.10.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
openai==0.27.7
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
paramiko==2.6.0
parso==0.5.2
pathspec==0.7.0
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==11.9.0
prompt-toolkit==2.0.10
protobuf==3.20.3
psutil==5.5.1
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2023.5.5
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
smmap2==2.0.5
sounddevice==0.4.6
stevedore==1.32.0
sympy==1.12
systemd-python==234
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
traitlets==4.3.3
transformers==4.29.2
triton==2.0.0
typing-extensions==4.6.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
wcwidth==0.1.8
xkit==0.0.0
yamllint==1.20.0
yarl==1.9.2
zipp==3.15.0
